--- a/.githooks/pre-commit.backup
+++ b/.githooks/pre-commit.backup
@@ -1,0 +1,42 @@
+#!/bin/bash
+set -e
+
+# Colors for output
+RED='\033[0;31m'
+GREEN='\033[0;32m'
+YELLOW='\033[1;33m'
+NC='\033[0m' # No Color
+
+echo "🔍 Running pre-commit checks..."
+
+# Check if we have staged Go files
+STAGED_GO_FILES=$(git diff --cached --name-only --diff-filter=ACM | grep '\.go$' || true)
+
+if [ -z "$STAGED_GO_FILES" ]; then
+    echo "No Go files staged for commit"
+    exit 0
+fi
+
+echo -e "\n${YELLOW}Running formatter checks...${NC}"
+make fmt-all
+
+echo -e "\n${YELLOW}Checking go.mod files...${NC}"
+make mod-tidy
+
+echo -e "\n${YELLOW}Running linter on all modules...${NC}"
+make lint-all
+
+echo -e "\n${YELLOW}Running tests on all modules...${NC}"
+make test-all
+
+# Check if any files were modified by formatting or go mod tidy
+MODIFIED=$(git diff --name-only)
+if [ -n "$MODIFIED" ]; then
+    echo -e "\n${RED}❌ The following files were modified:${NC}"
+    echo "$MODIFIED"
+    echo -e "${YELLOW}Please stage these changes and commit again.${NC}"
+    echo -e "${YELLOW}Run 'git add -u' to stage the changes.${NC}"
+    exit 1
+fi
+
+echo -e "\n${GREEN}✅ All pre-commit checks passed!${NC}"