--- a/core/test_precommit.go
+++ b/core/test_precommit.go
@@ -1,6 +1,0 @@
-package core
-
-// TestFunction is a test for the pre-commit hook
-func TestFunction() string {
-	return "testing pre-commit"
-}