--- a/rulebooks/dnd5e/examples/README.md
+++ b/rulebooks/dnd5e/examples/README.md
@@ -1,0 +1,16 @@
+# D&D 5e RPG-Toolkit Examples
+
+This directory contains example implementations showing how to integrate the rpg-toolkit D&D 5e calculations with rpg-api data structures.
+
+## Contents
+
+- `character_adapter.go` - Adapters to bridge rpg-api data structures with toolkit interfaces
+- `integration_test.go` - Comprehensive integration tests demonstrating real-world usage
+- `demo/main.go` - Interactive demo showing character calculations
+
+## Usage
+
+The examples demonstrate:
+1. How to adapt your existing data structures to work with the toolkit
+2. How to use the calculation engine for game mechanics
+3. Best practices for integrating with game state management