… rulebooks/dnd5e/PRECOMMIT_OPTIMIZATION.md